requirements.txt
alembic==1.7.3
autopep8==1.5.7
click==8.0.1
colorama==0.4.4
Flask==2.0.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
pycodestyle==2.7.0
python-dotenv==0.19.0
SQLAlchemy==1.4.25
toml==0.10.2
Werkzeug==2.0.1
